Move TxOut to primitives
The `TxOut` has all public fields; move the `TxOut` to `primitives`.
