+++ primitives/src/transaction.rs
@@ -18,7 +18,7 @@ use hashes::sha256d;
 #[cfg(feature = "alloc")]
 use internals::write_err;
 #[cfg(feature = "alloc")]
-use units::parse;
+use units::{parse, Amount};
 
 #[cfg(feature = "alloc")]
 use crate::script::ScriptBuf;
@@ -67,6 +67,34 @@ impl TxIn {
         sequence: Sequence::MAX,
         witness: Witness::new(),
     };
+}
+
+/// Bitcoin transaction output.
+///
+/// Defines new coins to be created as a result of the transaction,
+/// along with spending conditions ("script", aka "output script"),
+/// which an input spending it must satisfy.
+///
+/// An output that is not yet spent by an input is called Unspent Transaction Output ("UTXO").
+///
+/// ### Bitcoin Core References
+///
+/// * [CTxOut definition](https://github.com/bitcoin/bitcoin/blob/345457b542b6a980ccfbc868af0970a6f91d1b82/src/primitives/transaction.h#L148)
+#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Debug, Hash)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg(feature = "alloc")]
+pub struct TxOut {
+    /// The value of the output, in satoshis.
+    pub value: Amount,
+    /// The script which must be satisfied for the output to be spent.
+    pub script_pubkey: ScriptBuf,
+}
+
+#[cfg(feature = "alloc")]
+impl TxOut {
+    /// This is used as a "null txout" in consensus signing code.
+    pub const NULL: Self =
+        TxOut { value: Amount::from_sat(0xffffffffffffffff), script_pubkey: ScriptBuf::new() };
 }
 
 /// A reference to a transaction output.
@@ -258,6 +286,14 @@ impl<'a> Arbitrary<'a> for TxIn {
             sequence: Sequence::arbitrary(u)?,
             witness: Witness::arbitrary(u)?,
         })
+    }
+}
+
+#[cfg(feature = "arbitrary")]
+#[cfg(feature = "alloc")]
+impl<'a> Arbitrary<'a> for TxOut {
+    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
+        Ok(TxOut { value: Amount::arbitrary(u)?, script_pubkey: ScriptBuf::arbitrary(u)? })
     }
 }
